implemented full sdk android and ios improvements

+++ yuno_sdk_platform_interface/lib/src/features/show_payment_methods/yuno_payment_method_channel.dart
@@ -30,7 +30,7 @@ final class YunoPaymentMethodChannel implements YunoPaymentMethodPlatform {
         case 'onHeightChange':
           if (call.arguments is! double) return;
           final height = call.arguments as double;
-          controller.updateHeight(height);
+          controller.updateHeight(height.toDouble());
           break;
 
         case 'onSelected':
